// src/modules/users/interfaces/users.interface.ts
export enum UserRole {
  SUPER_ADMIN = 'super_admin',
  ADMIN = 'admin',
  USER = 'user',
}

export interface IUser {
  id: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  phoneCountryCode: string;
  country: string;
  city: string;
  email: string;
  role: UserRole;
  password: string;
  // dateOfBith: Date;
}

export interface IUserFilter {
  firstName?: string;
  lastName?: string;
  email?: string;
  role?: UserRole;
  country?: string;
  city?: string;
}